Added a method to the Meta baseclass called get_max_paramters()
It returns the maximum number of allowed parameters for that plugin. By default it returns None, which means there is no maximum. This also means, that plugins don't have to overwrite this method which won't break older plugins

As a result of this, the bot itself can check if the command is even possible for the plugin before delegating it to said plugin. One thing less to check on the plugin side and a unified error message for that